pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "film-intelligence-agent"
version = "0.1.0"
description = "Production-ready film intelligence agent for Lodewijk Vos."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.14.1",
  "beautifulsoup4>=4.12.3",
  "httpx>=0.28.1",
  "jinja2>=3.1.6",
  "lxml>=5.3.2",
  "notion-client>=2.3.0",
  "pynacl>=1.5.0",
  "orjson>=3.10.16",
  "psycopg[binary]>=3.2.6",
  "pydantic>=2.10.6",
  "pydantic-settings>=2.8.1",
  "python-dateutil>=2.9.0.post0",
  "python-dotenv>=1.0.1",
  "pyyaml>=6.0.2",
  "resend>=2.7.0",
  "sqlalchemy>=2.0.39",
  "tenacity>=9.0.0",
  "typer>=0.15.2",
]

[project.optional-dependencies]
dev = [
  "freezegun>=1.5.1",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.26.0",
  "ruff>=0.11.4",
]

[project.scripts]
film-intel = "film_intel_agent.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
